requirements.txt
vosk
sounddevice
pyttsx3
scikit-learn
requests
pyyaml
RPi.GPIO
